Refactoring question app and create grat app
