+++ pgtbl/grat/models/__init__.py
@@ -1,0 +1,1 @@
+from .grat_submission import GRATSubmission
