+++ pgtbl/grat/models/grat_submission.py
@@ -3,11 +3,10 @@ from django.db import models
 from django.conf import settings
 from groups.models import Group
 from modules.models import TBLSession
-from .question import Question
-from .submission import Submission
+from questions.models.question import Question
 
 
-class GRATSubmission(Submission):
+class GRATSubmission(models.Model):
     """
     Store all submissions for a given gRAT question.
     """
@@ -38,6 +37,24 @@ class GRATSubmission(Submission):
         on_delete=models.CASCADE,
         verbose_name=_('Groups'),
         related_name="submissions"
+    )
+
+    correct_alternative = models.CharField(
+        _('Correct Alternative'),
+        max_length=1000,
+        help_text=_('Correct alternative title.')
+    )
+
+    score = models.PositiveIntegerField(
+        _("Score"),
+        default=0,
+        help_text=_("Question score answered."),
+    )
+
+    created_at = models.DateTimeField(
+        _('Created at'),
+        help_text=_("Date that the submission of question is created."),
+        auto_now_add=True
     )
 
     def __str__(self):
