+++ pgtbl/grat/templates/grat/progress_bar.html
