+++ pgtbl/grat/templates/grat/result.html
@@ -4,13 +4,13 @@
 
 {% block css %}
   <link rel="stylesheet" href="{% static 'accounts/css/sidebar.css' %}" />
-  <link rel="stylesheet" href="{% static 'questions/css/result.css' %}" />
+  <link rel="stylesheet" href="{% static 'grat/css/result.css' %}" />
 {% endblock %}
 
 {% block breadcrumb %}
   {{ block.super }}
   <li>
-    <a href="{% url 'questions:grat-result' discipline.slug session.id %}">
+    <a href="{% url 'grat:result' discipline.slug session.id %}">
        {% trans 'gRAT result' %}
     </a>
   </li>
