+++ pgtbl/grat/templates/grat/update.html
@@ -10,7 +10,7 @@
         </h4>
       </div>
       <div class="modal-body">
-        <form action="{% url 'questions:grat-update' discipline.slug session.id %}" method="post" novalidate>
+        <form action="{% url 'grat:update' discipline.slug session.id %}" method="post" novalidate>
           <!-- Form security -->
           {% csrf_token %}
 
@@ -23,10 +23,10 @@
             <input
               type="number"
               class="form-control"
-              id="{{grat_form.weight.id_for_label}}"
+              id="{{grat_form.grat_weight.id_for_label}}"
               min="0"
-              name="{{grat_form.weight.name}}"
-              value="0"
+              name="{{grat_form.grat_weight.name}}"
+              value="{{ session.grat_weight }}"
             />
           </div>
