+++ pgtbl/grat/views/__init__.py
@@ -1,0 +1,5 @@
+from .grat_answer_question_view import GRATAnswerQuestionView
+from .grat_date_update_view import GRATDateUpdateView
+from .grat_result_view import GRATResultView
+from .grat_update_view import GRATUpdateView
+from .grat_view import GRATView
