+++ pgtbl/modules/templates/modules/sidebar.html
@@ -65,7 +65,7 @@
 
     <!-- Session gRAT menu -->
     {% if date > grat_datetime and user in discipline.students.all %}
-      <a href="{% url 'questions:grat-result' discipline.slug session.id %}"
+      <a href="{% url 'grat:result' discipline.slug session.id %}"
          class="btn"
          data-toggle="tooltip"
          data-placement="right"
@@ -73,7 +73,7 @@
          <i class="fa fa-group icon-lg"></i>
       </a>
     {% else %}
-      <a href="{% url 'questions:grat-list' discipline.slug session.id %}"
+      <a href="{% url 'grat:list' discipline.slug session.id %}"
          class="btn"
          data-toggle="tooltip"
          data-placement="right"
