+++ pgtbl/questions/forms/__init__.py
@@ -1,3 +1,2 @@
 from .alternative_form import AlternativeForm, AlternativeFormSet
 from .question_form import QuestionForm, AnswerQuestionForm
-from .grat_form import GRATForm, GRATDateForm, AnswerGRATQuestionForm
