+++ pgtbl/questions/models/__init__.py
@@ -1,3 +1,2 @@
 from .alternative import Alternative
 from .question import Question
-from .grat_submission import GRATSubmission
