+++ pgtbl/questions/urls.py
@@ -21,41 +21,9 @@ questions_patterns = [
     ),
 ]
 
-grat_patterns = [
-    path(
-        '',
-        views.GRATView.as_view(),
-        name='grat-list'
-    ),
-    path(
-        'edit-date/',
-        views.GRATDateUpdateView.as_view(),
-        name='grat-date'
-    ),
-    path(
-        'edit-irat/',
-        views.GRATUpdateView.as_view(),
-        name='grat-update'
-    ),
-    path(
-        'result/',
-        views.GRATResultView.as_view(),
-        name='grat-result'
-    ),
-    path(
-        'question/<int:question_id>/answer-page/<int:question_page>/',
-        views.GRATAnswerQuestionView.as_view(),
-        name='grat-answer-question'
-    ),
-]
-
 urlpatterns = [
     path(
         'profile/<slug:slug>/sessions/<int:pk>/questions/',
         include(questions_patterns)
-    ),
-    path(
-        'profile/<slug:slug>/sessions/<int:pk>/grat/',
-        include(grat_patterns)
-    ),
+    )
 ]
