+++ pgtbl/questions/views/__init__.py
@@ -1,8 +1,3 @@
-from .grat.grat_answer_question_view import GRATAnswerQuestionView
-from .grat.grat_date_update_view import GRATDateUpdateView
-from .grat.grat_result_view import GRATResultView
-from .grat.grat_update_view import GRATUpdateView
-from .grat.grat_view import GRATView
-from .question.create_question_view import CreateQuestionView
-from .question.delete_question_view import DeleteQuestionView
-from .question.update_question_view import UpdateQuestionView
+from .create_question_view import CreateQuestionView
+from .delete_question_view import DeleteQuestionView
+from .update_question_view import UpdateQuestionView
