+++ pgtbl/tbl/urls.py
@@ -17,6 +17,7 @@ urlpatterns = [
     url(r'', include('practical_test.urls')),
     url(r'', include('exercises.urls')),
     url(r'', include('irat.urls')),
+    url(r'', include('grat.urls'))
 ]
 
 # While in development mode we will use relative URL for static and average
